+++ internal/sender/service.go
@@ -14,6 +14,8 @@ import (
 	"firebase.google.com/go/v4/messaging"
 	"github.com/google/uuid"
 	coresdk "github.com/goverland-labs/core-web-sdk"
+	"github.com/goverland-labs/core-web-sdk/dao"
+	"github.com/goverland-labs/core-web-sdk/proposal"
 	"github.com/goverland-labs/inbox-api/protobuf/inboxapi"
 	"github.com/rs/zerolog/log"
 	"google.golang.org/api/option"
@@ -35,6 +37,27 @@ type UsersFinder interface {
 
 type SettingsProvider interface {
 	GetPushDetails(ctx context.Context, in *inboxapi.GetPushDetailsRequest, opts ...grpc.CallOption) (*inboxapi.GetPushDetailsResponse, error)
+	GetPushToken(ctx context.Context, in *inboxapi.GetPushTokenRequest, opts ...grpc.CallOption) (*inboxapi.PushTokenResponse, error)
+	GetPushTokenList(ctx context.Context, in *inboxapi.GetPushTokenListRequest, opts ...grpc.CallOption) (*inboxapi.PushTokenListResponse, error)
+}
+
+type MessageSender interface {
+	Send(ctx context.Context, message *messaging.Message) (string, error)
+}
+
+type CoreDataProvider interface {
+	GetUserVotes(ctx context.Context, address string, params coresdk.GetUserVotesRequest) (*proposal.VoteList, error)
+	GetDao(ctx context.Context, id uuid.UUID) (*dao.Dao, error)
+	GetProposal(ctx context.Context, id string) (*proposal.Proposal, error)
+}
+
+type DataManipulator interface {
+	Create(item *History) error
+	GetByHash(hash string) (*History, error)
+	MarkAsClicked(messageUUID uuid.UUID) error
+	QueueByFilters(_ context.Context, filters []Filter) ([]SendQueue, error)
+	CreateSendQueueRequest(_ context.Context, item *SendQueue) error
+	MarkAsSent(_ context.Context, ids []uint) error
 }
 
 type cacheItem struct {
@@ -47,8 +70,8 @@ type Service struct {
 	subscriptions SubscriptionsFinder
 	usrs          UsersFinder
 	settings      SettingsProvider
-	client        inboxapi.SettingsClient
-	core          *coresdk.Client
+	core          CoreDataProvider
+	sender        MessageSender
 
 	cache map[string]cacheItem
 	mu    sync.Mutex
@@ -60,7 +83,6 @@ type Service struct {
 func NewService(
 	r *Repo,
 	cfg config.Push,
-	client inboxapi.SettingsClient,
 	subs SubscriptionsFinder,
 	usrs UsersFinder,
 	sp SettingsProvider,
@@ -71,21 +93,27 @@ func NewService(
 		return nil, err
 	}
 
+	// todo: check if it can live days...
+	sender, err := makeSender(context.Background(), data, cfg.ProjectID)
+	if err != nil {
+		return nil, fmt.Errorf("failed to make sender: %w", err)
+	}
+
 	return &Service{
 		repo:          r,
 		subscriptions: subs,
 		usrs:          usrs,
 		settings:      sp,
-		client:        client,
 		cfg:           data,
 		projectID:     cfg.ProjectID,
+		sender:        sender,
 		core:          coreSDK,
 		cache:         make(map[string]cacheItem),
 	}, nil
 }
 
 func (s *Service) GetToken(ctx context.Context, userID uuid.UUID) (string, error) {
-	response, err := s.client.GetPushToken(ctx, &inboxapi.GetPushTokenRequest{UserId: userID.String()})
+	response, err := s.settings.GetPushToken(ctx, &inboxapi.GetPushTokenRequest{UserId: userID.String()})
 	if err != nil {
 		return "", fmt.Errorf("get push token by user_id: %s: %w", userID, err)
 	}
@@ -94,7 +122,7 @@ func (s *Service) GetToken(ctx context.Context, userID uuid.UUID) (string, error
 }
 
 func (s *Service) GetTokens(ctx context.Context, userID uuid.UUID) ([]TokenDetails, error) {
-	response, err := s.client.GetPushTokenList(ctx, &inboxapi.GetPushTokenListRequest{UserId: userID.String()})
+	response, err := s.settings.GetPushTokenList(ctx, &inboxapi.GetPushTokenListRequest{UserId: userID.String()})
 	if err != nil {
 		return nil, fmt.Errorf("get push tokens by user_id: %s: %w", userID, err)
 	}
@@ -124,130 +152,12 @@ func (r request) hash() string {
 	return hex.EncodeToString(hash[:])
 }
 
-// Deprecated: use SendV2 function instead, will be removed in next releases
 func (s *Service) Send(ctx context.Context, req request) error {
-	item, err := s.repo.GetByHash(req.hash())
-	if err != nil && !errors.Is(err, gorm.ErrRecordNotFound) {
-		return fmt.Errorf("getByHash: %w", err)
-	}
-	if item != nil {
-		log.Warn().Msgf("duplicate sending push: %s %s", req.userID.String(), req.title)
-
-		return nil
-	}
-
-	last, err := s.repo.GetLastSent(req.userID)
-	if err != nil && !errors.Is(err, gorm.ErrRecordNotFound) {
-		return fmt.Errorf("repo.GetLastSent: %w", err)
-	}
-
-	// to avoid spamming
-	if last != nil && time.Since(last.CreatedAt) <= time.Minute {
-		log.Warn().Msgf("sending less than in minute: %s %s %v", req.userID.String(), req.title, time.Since(last.CreatedAt).Seconds())
-
-		return nil
-	}
-
-	client, err := s.makeClient(ctx)
-	if err != nil {
-		return fmt.Errorf("s.makeClient: %w", err)
-	}
-
-	response, err := client.Send(ctx, &messaging.Message{
-		Notification: &messaging.Notification{
-			Title:    req.title,
-			Body:     req.body,
-			ImageURL: req.imageURL,
-		},
-		Token: req.token,
-	})
-	if err != nil {
-		log.Error().Err(err).Msg("send push")
-
-		return nil
-	}
-
-	if err = s.repo.Create(&History{
-		UserID: req.userID,
-		Message: Message{
-			Title:    req.title,
-			Body:     req.body,
-			ImageURL: req.token,
-		},
-		PushResponse: response,
-		Hash:         req.hash(),
-	}); err != nil {
-		log.Error().Err(err).Msg("create history log")
-	}
-
-	return nil
-}
-
-// Deprecated: use SendV2 function instead, will be removed in next releases
-func (s *Service) SendCustom(ctx context.Context, req request) error {
-	client, err := s.makeClient(ctx)
-	if err != nil {
-		return fmt.Errorf("s.makeClient: %w", err)
-	}
-
-	msgID := uuid.New()
-	response, err := client.Send(ctx, &messaging.Message{
-		Token: req.token,
-		Notification: &messaging.Notification{
-			Title:    req.title,
-			Body:     req.body,
-			ImageURL: req.imageURL,
-		},
-		APNS: &messaging.APNSConfig{
-			Payload: &messaging.APNSPayload{
-				Aps: &messaging.Aps{
-					MutableContent: true,
-				},
-				CustomData: map[string]interface{}{
-					"id":        msgID,
-					"proposals": req.payload,
-				},
-			},
-			FCMOptions: &messaging.APNSFCMOptions{
-				ImageURL: req.imageURL,
-			},
-		},
-	})
-	if err != nil {
-		log.Error().Err(err).Msg("send custom push")
-
-		return nil
-	}
-
-	if err = s.repo.Create(&History{
-		UserID: req.userID,
-		Message: Message{
-			ID:       msgID,
-			Title:    req.title,
-			Body:     req.body,
-			ImageURL: req.token,
-			Payload:  req.payload,
-		},
-		PushResponse: response,
-		Hash:         uuid.NewString(),
-	}); err != nil {
-		log.Error().Err(err).Msg("create history log")
-	}
-
-	return nil
-}
-
-func (s *Service) SendV2(ctx context.Context, req request) error {
 	list, err := s.GetTokens(context.TODO(), req.userID)
 	if err != nil {
 		log.Warn().Err(err).Msgf("get token for user %s", req.userID.String())
 
 		return nil
-	}
-
-	client, err := s.makeClient(ctx)
-	if err != nil {
-		return fmt.Errorf("s.makeClient: %w", err)
 	}
 
 	msgID := uuid.New()
@@ -263,7 +173,7 @@ func (s *Service) SendV2(ctx context.Context, req request) error {
 			continue
 		}
 
-		response, err := client.Send(ctx, &messaging.Message{
+		response, err := s.sender.Send(ctx, &messaging.Message{
 			Token: info.Token,
 			Notification: &messaging.Notification{
 				Title:    req.title,
@@ -315,10 +225,10 @@ func (s *Service) SendV2(ctx context.Context, req request) error {
 	return nil
 }
 
-func (s *Service) makeClient(ctx context.Context) (*messaging.Client, error) {
-	authOpt := option.WithCredentialsJSON(s.cfg)
+func makeSender(ctx context.Context, cfg []byte, projectID string) (MessageSender, error) {
+	authOpt := option.WithCredentialsJSON(cfg)
 	fapp, err := firebase.NewApp(context.Background(), &firebase.Config{
-		ProjectID: s.projectID,
+		ProjectID: projectID,
 	}, authOpt)
 	if err != nil {
 		return nil, fmt.Errorf("create firebase app: %w", err)