+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+- Basic unit tests
+
 ### Changed
 - Use push settings details to avoid sending irrelevant notifications