+++ CHANGELOG.md
@@ -7,11 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.1.1] - 2024-07-30
+
 ### Added
 - Basic unit tests
-
-### Changed
-- Use push settings details to avoid sending irrelevant notifications
 
 ## [0.1.0] - 2024-07-30